Chaging saving unix timestamp from double to long

+++ GazeMonitoring.Model/GazePoint.cs
@@ -2,6 +2,6 @@
     public class GazePoint {
         public double X { get; set; }
         public double Y { get; set; }
-        public double Timestamp { get; set; }
+        public long Timestamp { get; set; }
     }
 }
